services/dashboard/requirements.txt
streamlit>=1.32.0
httpx>=0.27.0
pandas>=2.2.0
psycopg2-binary>=2.9.9
pgvector>=0.2.4
plotly>=5.19.0
